Change to webpack build
Previously this module was designed as a jspm module for working with
Bigcommerce themes. However, having it available as a generic npm
module ensures that it's more flexible and reusable with future
projects. Because jspm has been removed as a dependency, we need
to ensure the build process bundles without requiring Systemjs.

- Move some files around to keep project organized: source files "src"
- Move all the demo files into one directory
- Set up and install webpack with a configuration for building the
  module, as well as a configuration for running the demo locally
- Update the dependencies/package/readme
- Start a changelog so we have a record moving foward of changes
  applied to the module